Move Piece class to separate module
