+++ src/main.js
@@ -17,7 +17,7 @@ limitations under the License.
 
 module.exports = {
     board: require('./board'),
-    pieces: require('./pieces'),
+    pieces: require('./pieces/main'),
     relations: require('./relations'),
     square: require('./square')
 };
