+++ src/pieces/main.js
@@ -15,256 +15,9 @@ limitations under the License.
 */
 
 
-const { Relation, PieceSquares } = require('./relations');
-const { SquareCoordinates } = require('./square');
-
-
-class Piece {
-    // Base chess piece class.
-
-    static WHITE = 'white';
-    static BLACK = 'black';
-
-    static PAWN = 'pawn';
-    static KNIGHT = 'knight';
-    static BISHOP = 'bishop';
-    static ROOK = 'rook';
-    static QUEEN = 'queen';
-    static KING = 'king';
-
-    static ALL_KINDS = [Piece.PAWN, Piece.KNIGHT, Piece.BISHOP, Piece.ROOK, Piece.QUEEN, Piece.KING];
-    static ALL_COLORS = [Piece.WHITE, Piece.BLACK];
-    static ALL_LINEARS = [Piece.BISHOP, Piece.ROOK, Piece.QUEEN];
-
-    constructor(color, square, kind=null, refresh=true) {
-        /*
-        Params:
-            color {string} (white or black);
-            square {Square} - where piece is placed;
-            kind {string} - one of Piece.ALL_KINDS;
-            refresh {boolean} - whether refresh board after piece placed or not.
-        */
-
-        this._setKind(kind);
-        this._setColor(color);
-        this._isLinear = Piece.ALL_LINEARS.includes(kind);
-        this._getKind();
-        this.squares = new PieceSquares(this);
-        this._refreshSquareFinder();
-        this.getInitState();
-        this.getPlace(square, refresh);
-    }
-
-    get color() {
-        return this._color;
-    }
-
-    get stuck() {
-        // check the piece get stucked
-        return !this.squares[Relation.MOVE] && !this.squares[Relation.ATTACK];
-    }
-
-    get kind() {
-        return this._kind;
-    }
-
-    get isLinear() {
-        return this._isLinear;
-    }
-
-    get board() {
-        return this.square.board;
-    }
-
-    _setKind(kind) {
-        /*
-        Params:
-            kind {string} - one of Piece.ALL_KINDS.
-        */
-
-        if (kind != null && !Piece.ALL_KINDS.includes(kind)) {
-            throw Error(`'${kind}' is wrong piece kind value. Use any of Piece.ALL_KINDS.`);
-        }
-        this._kind = kind;
-    }
-
-    _setColor(color) {
-        /*
-        Params:
-            color {string} - one of Piece.ALL_COLORS.
-        */
-
-        if (!Piece.ALL_COLORS.includes(color)) {
-            throw Error(`'${color}' is wrong piece color value. Use any of Piece.ALL_COLORS.`);
-        }
-        this._color = color;
-    }
-
-    _refreshSquareFinder() {
-        // Refresh square finder states
-
-        this.sqrBeforeXray = null; // square before xray (always occupied by piece)
-        this.xrayControl = false; // control square behind checked king (inline of piece attack)
-        this.endOfALine = false;
-    }
-
-    _refreshSquares() {
-        // Refresh piece squares data
-
-        this.squares.refresh();
-        this._refreshSquareFinder();
-    }
-
-    _getKind() {
-        this.isPawn = this.kind == Piece.PAWN;
-        this.isKnight = this.kind == Piece.KNIGHT;
-        this.isBishop = this.kind == Piece.BISHOP;
-        this.isRook = this.kind == Piece.ROOK;
-        this.isQueen = this.kind == Piece.QUEEN;
-        this.isKing = this.kind == Piece.KING;
-    }
-
-    _nextSquareAction(nextSquare) {
-        /*
-        Handle action with next square.
-        Params:
-            nextSquare {Square}.
-        */
-
-        if (this.sqrBeforeXray) {
-            this.squares.add(Relation.XRAY, nextSquare);
-            if (this.xrayControl) {
-                this.squares.add(Relation.CONTROL, nextSquare);
-                this.xrayControl = false;
-            }
-            if (nextSquare.piece) {
-                let isOppKingSquare = nextSquare.piece.isKing && !this.sameColor(nextSquare.piece);
-                let isOppPieceBeforeXray = !this.sameColor(this.sqrBeforeXray.piece);
-                if (isOppKingSquare && isOppPieceBeforeXray) {
-                    this.sqrBeforeXray.piece.binder = this;
-                }
-                this.endOfALine = true;
-            }
-        }
-        else if (nextSquare.piece) {
-            if (this.sameColor(nextSquare.piece)) {
-                this.squares.add(Relation.COVER, nextSquare);
-            }
-            else {
-                this.squares.add(Relation.ATTACK, nextSquare);
-                if (nextSquare.piece.isKing) {
-                    nextSquare.piece.checkers.add(this.square.piece);
-                    this.xrayControl = true;
-                }
-            }
-            this.squares.add(Relation.CONTROL, nextSquare);
-            if (this.isLinear) this.sqrBeforeXray = nextSquare;
-        }
-        else {
-            this.squares.add(Relation.MOVE, nextSquare);
-            this.squares.add(Relation.CONTROL, nextSquare);
-        }
-    }
-
-    getSquares() {
-        // Get piece squares by piece action
-    }
-
-    getInitState() {
-        this.binder = null;
-    }
-
-    getPlace(square, refresh=true) {
-        /*
-        Place piece to square.
-        Params:
-            square {Square};
-            refresh {boolean} - whether refresh board after piece placed or not.
-        */
-
-        this.square = square;
-        square.placePiece(this, refresh);
-    }
-
-    theSame(otherPiece) {
-        /*
-        Check piece is the same with other piece.
-        Params:
-            otherPiece {Piece / its subclass}.
-        */
-
-        return this.square.theSame(otherPiece.square);
-    }
-
-    sameColor(otherPiece) {
-        /*
-        Check piece has same color with other piece.
-        Params:
-            otherPiece {Piece / its subclass}.
-        */
-
-        return this.color === otherPiece.color;
-    }
-
-    hasColor(color) {
-        /*
-        Check piece has particular color.
-        Params:
-            color {string} (one of Piece.ALL_COLORS).
-        */
-
-        return this.color === color;
-    }
-
-    canBeReplacedTo(square) {
-        /*
-        Check piece could be replaced to particular square.
-        Params:
-            square {Square}.
-        */
-
-        return (
-            !square.piece
-            &&
-            this.squares.includes(Relation.MOVE, square)
-        ||
-            !!square.piece
-            &&
-            !square.piece.isKing
-            &&
-            !this.sameColor(square.piece)
-            &&
-            this.squares.includes(Relation.ATTACK, square)
-        );
-    }
-
-    getTotalImmobilize() {
-        // Immobilize piece.
-
-        for (let kind of [Relation.MOVE, Relation.ATTACK, Relation.COVER]) {
-            this.squares.refresh(kind);
-        }
-    }
-
-    getBind(kingSquare) {
-        // Bind piece.
-
-        this.squares.refresh(Relation.XRAY);
-        let betweenSquares = this.binder.square.getBetweenSquaresNames(kingSquare, true, true);
-        for (let actonKind of [Relation.MOVE, Relation.ATTACK, Relation.COVER]) {
-            this.squares.limit(actonKind, betweenSquares);
-        }
-    }
-
-    getCheck(checker, betweenSquaresNames) {
-        // Change piece action abilities after its king was checked
-
-        this.squares.refresh(Relation.COVER);
-        this.squares.refresh(Relation.XRAY);
-        this.squares.limit(Relation.ATTACK, [checker.square.name.value]);
-        this.squares.limit(Relation.MOVE, betweenSquaresNames);
-    }
-}
+const { Piece } = require('./base');
+const { Relation } = require('../relations');
+const { SquareCoordinates } = require('../square');
 
 
 class Pawn extends Piece {
@@ -977,7 +730,7 @@ class King extends StepPiece {
         }
     }
 
-    getInitState() {
+    setInitState() {
         this.checkers = new KingCheckers(this);
     }
